Adding the network module and trainer structure
Start the necessary module for building the TensorMol network. TensorMol class contains the network weights and biases, accepts the featurized inputs, and returns the latent featurization from the final layer.

Subnetworks classes are implemented to accommodate unique elements for each subnetwork. TensorMol will build the number of subnetworks necessary based on the data it is to be trained on.

The TensorMol trainer class handles coordinating the dataset, the featurization, the network, and the optimizers.