building the radial symmetry functions
Primarily implements the calculation of radial symmetry functions in a dense way. Still needs the inclusion of a neighbor list to build the angular symmetry functions (from ASE or Jax-MD) in a sparse manner. Testing needs to be added as well.

Data objects have been updated. Geometry objects now have an atomic_nums property because it can otherwise be difficult to feed the atomic number from the Geometry object itself. Some work on how the Geometry and MoleculeSet objects interact would be good to smooth things out here.

Dataset objects now save their own JSON filetype with a “.dset” extension. Even though this can contain the same data as the MoleculeSet file, the reason this could be convenient is to prepare a dataset on a storage server full of msets, and then transfer the dataset to a compute server for training or inference.

Other changes include rebranding the package to TensorChem, Code style fixes, and a few bug fixes.

remove unneeded import

Small fixes