Reworking geometry property names a little
Renaming properties to labels because it is shorter to type. Also starting an atom object.

Additionally this changed the make_dataset to use the JSON library directly, which should simplify thing a bit.

Fixes to correct for the tests

Fixes to pass the tests! Alexis your tests are already saving us from me making a mess!

Adding first example

Adding a first example iPython notebook