Modifying MSets and Converting to SMILES
Focusing on the AIMD and metadynamics data, MSets were created and modified. These MSets were converted into SMILES with a smiles converter which was edited to only convert the minimum geometry into a SMILES string with correct charge and minimal errors through the conversion. The SMILES of the AIMD and meta MSets were compared with a comparison script. The ANI-1x data was converted into MSets with smiles identifiers and the GDB-9 data was identified with SMILES as well. Other changes that were made include updating the MixedDataset to match new MSet format, creating a Jupyter Notebook which contains and displays information about the scope of the MSets, making xyz2mol a util submodule, and cleaning a few unneeded scripts.