Remakes scripts for crawling Q-Chem output files
Adds a remade script which should import an AIMD trajectory from a Q-Chem file better. It should be more robust in dealing with errors and exiting more gracefully if the file is incomplete. It builds the MSet iteratively as it goes instead of building it all once all of the data has been collected. At each step it attempts to add a new geometry to the MSet if it collected a new sample of data over the previous step. If it succeeds, it continues trying to collect more steps until it reaches the end of the file. If it fails, it returns the MSet at that point, so that the geometries which were successfully collected will still be returned. Could use some better exception handling still, but has only been debugged on one good file so far.

fix to AIMD crawler script

Adds script similar to AIMD crawler for metadynamics

Adds another script similar to the AIMD crawler but this script works for metadynamics trajectories which are run as multiple single point calculations in one file. The method needs changed slightly to be adapted to work for mutiple files for a single MSet like the format of the less50 data is currently in.

updates to metadynamics crawler script

Adds another metadynamics crawling script

This one works on directories of outputs which contain a single calculation in each file, but all subdirectories are of the same molecule

fixes to get multifile sp working