Adding labels classes
This adds objects for holding labels, which should alleviate some of the confusion about what the keys should be. Labels objects have a basis and method and possibly partitioning scheme in the case of partial charges. Labels export their own json, so atoms and geometries need just call the export_json() method for each label they own in their labels dictionary and add the result to their own json for exporting.

Also modularizes the q-chem crawling scripts to make them easier to maintain, and adds the labels objects to them.